refactor(mcp): centralize model-visible Apps file schemas
Move Codex Apps file-param schema masking onto ToolInfo so direct MCP exposure and tool_search use the same model-visible projection.

Keep execution-time upload argument rewriting at the MCP tool-call boundary while sharing the file-param metadata parser from codex-mcp.


docs: update AGENTS.md
