+++ codex-rs/core/src/codex.rs
@@ -273,7 +273,6 @@ use crate::injection::app_id_from_path;
 use crate::injection::tool_kind_for_path;
 use crate::instructions::UserInstructions;
 use crate::mcp::McpManager;
-use crate::mcp_openai_file::mask_model_visible_tool_input_schema;
 use crate::mcp_skill_dependencies::maybe_prompt_and_install_mcp_dependencies;
 use crate::memories;
 use crate::mentions::build_connector_slug_counts;
